Deferred construction everywhere + more composition with UDA-based traits (WiP)
- Less is more. UDAs >> DbI. Pipeline composition now uses only UDAs to
make decisions about how things should be composed. This makes the code much
more concise, which much less corner cases to care about.
- Always use deferred composition. Pipeline is now a tree with the root
being all-active stage pulling data from its source (if any) and pushing it
to its sink (if any). Now stages can be non-copyable and can use internal
pointers (a struct should not be moved while a member func of it or of its
member is running, right?). Args to stage ctors must be copyable. A simple
trick to avoid additional indirections is to actually acquire the managed
resource only in stage ctor or later, and assert that any copies of the
object are made before that.
- All passive vs. active permutations are working with this approach.
Passive-to-passive is now a single object and does not need to use fibers
explicitly, just mixes in a Scheduler and calls yield().
Arbitrarily long active source chains can be wrapped in passive source
interface (it might be challenging for optimizers, though).
- Still missing: tests for alloc sink-passive source filters; implicit
adapters for incompatible stages.