[pipeline] Added support for source-driven pipelines.
Fibers are now spawned in Pipeline.run() instead of in the middle of
Pipeline instance construction.
In unit test mode there's a special call Schema.drive, which sets
the drive mode for the entire pipeline (i.e. whether it should be driven
by the leftmost (source) or rightmost (sink) driver).
In the future, this could be exposed in the form of a pipeline option
to allow overriding the default sink-driven behaviour in user code
if there are e.g. performance differences or just to test if a stage
runs properly in both modes (should only affect passive filters).