[pipeline] Insert adapters only when the pipeline is instantiated.
Now that a stage can implement multiple methods, we can't insert adapters
eagerly, because it is not known which methods will be used, and what kind
of adapter (if any) will be required, until the final sink is attached.
Stages are now appended to the schema without adapters, and only when
Schema.create is called (implicitly or explicitly), the exact methods
are chosen and any required adapters are inserted.