[pipeline] Refcounted ranges, forward put() to first stage.
Fibers are created with struct ptrs as delegate context ptrs, and these
pointers become invalid as soon as the struct is blitted to a different
address. Blitting won't happen for pipelines that are created in place
on the stack and executed immediately, but range objects are passed
around, which caused the fiber calls to reference the pipeline stages
via obsolete pointers.
The workaround is therefore to use a refcounted wrapper with
the actual pipeline object allocated on the heap using malloc.
If the additional indirection appears to impose significant performance
penalty, it should be possible to implement fibers in a way that does
not require storing context pointers in the fiber object (another
reason for reimplementing fibers is that they rely on GC).