[adapter] Do not throw OOMErrors
Semantics of Buffer.alloc() and Stage.alloc() is not scritly
defined yet.
Buffer should use bool alloc(ref T[], size_t)
so that e.g. FallbackBuffer is possible.
Stage should probably just throw when it's unable to fulfill
the request. Or maybe it should return false and give the
largest buffer it can at the moment, and let the source decide.